Role (#73)
* statistics changed

+++ src/main/java/uz/pdp/model/Rate.java
@@ -22,6 +22,5 @@ public class Rate {
     @ManyToOne
     private User user;
 
-    private Integer rate;
-
+    private Double rate;
 }
